+++ RELEASE-NOTES.md
@@ -1,5 +1,8 @@
 These are the release notes for the [MediaWiki][mediawiki] [GraphViz extension][gv_ext].
 
+## GraphViz 1.5.x ## (not released yet)
+* Add tag arguments preparse="dynamic" and preparse="static".
+
 ## GraphViz 1.4.1 ## (2014-10-21)
 * Fix for [bug 72325](https://bugzilla.wikimedia.org/show_bug.cgi?id=72325).